banner: fix dispose, refactor, add orientation handle, use default size if null, check mount

+++ lib/src/banner/adaptive/good_banner_adaptive_anchored.dart
@@ -1,6 +1,5 @@
 import 'package:common/common.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_good_ads/src/local_storage.dart';
 import 'package:google_mobile_ads/google_mobile_ads.dart';
 
 class GoodBannerAdaptiveAnchored extends StatefulWidget {
@@ -8,7 +7,6 @@ class GoodBannerAdaptiveAnchored extends StatefulWidget {
     Key? key,
     required this.adUnitId,
     this.adRequest = const AdRequest(),
-    this.interval = 60000,
     this.adsPlaceholderColor,
     this.onLoadAdError,
     this.onAdImpression,
@@ -17,7 +15,6 @@ class GoodBannerAdaptiveAnchored extends StatefulWidget {
 
   final String? adUnitId;
   final AdRequest adRequest;
-  final int interval;
   final Color? adsPlaceholderColor;
   final void Function(Object, StackTrace)? onLoadAdError;
   final void Function(int time, String adUnitId)? onAdImpression;
@@ -31,19 +28,21 @@ class GoodBannerAdaptiveAnchored extends StatefulWidget {
 
 class _GoodBannerAdaptiveAnchoredState
     extends State<GoodBannerAdaptiveAnchored> {
-  BannerAd? _anchoredAdaptiveAd;
+  BannerAd? _bannerAd;
   bool _isLoaded = false;
-  AnchoredAdaptiveBannerAdSize? size;
-  Size? defaultBannerSize;
+  AdSize? size;
+  AdSize? defaultBannerSize;
   String? adUnitId;
+  late Orientation _currentOrientation;
 
   @override
   void didChangeDependencies() {
     super.didChangeDependencies();
+    _currentOrientation = MediaQuery.of(context).orientation;
     computeDefaultBannerSize();
     if (widget.adUnitId != null) {
       adUnitId = widget.adUnitId!;
-      _initAd();
+      _loadAd();
     } else {
       _hideAd();
     }
@@ -55,7 +54,7 @@ class _GoodBannerAdaptiveAnchoredState
     WidgetsBinding.instance.addPostFrameCallback((_) {
       if (widget.adUnitId != null) {
         adUnitId = widget.adUnitId!;
-        _initAd();
+        _loadAd();
       } else {
         _hideAd();
       }
@@ -66,65 +65,51 @@ class _GoodBannerAdaptiveAnchoredState
     final width = MediaQuery.of(context).size.width;
     final height = width * 0.1557;
     setState(() {
-      defaultBannerSize = Size(width, height);
+      defaultBannerSize = AdSize(width: width.toInt(), height: height.toInt());
     });
   }
 
   void _hideAd() {
     setState(() {
-      _anchoredAdaptiveAd = null;
       _isLoaded = false;
     });
   }
 
-  Future<void> _initAd() async {
-    final lastImpressions = await getLastImpressions(adUnitId!);
-    if (DateTime.now().millisecondsSinceEpoch - lastImpressions >
-        widget.interval) {
-      _loadAd();
-    } else {
-      // if initState but can not load immediately, delay and then load
-      Future.delayed(Duration(milliseconds: widget.interval), () {
-        if (_isLoaded == false) {
-          _loadAd();
-        }
-      });
-    }
-  }
-
   Future<void> _loadAd() async {
+    if (!mounted) return;
     try {
       // Get an AnchoredAdaptiveBannerAdSize before loading the ad.
       size = await AdSize.getCurrentOrientationAnchoredAdaptiveBannerAdSize(
           MediaQuery.of(context).size.width.truncate());
 
       if (size == null) {
-        printDebug('Unable to get height of anchored banner.');
-        return;
+        printDebug('Unable to get height of anchored banner. use default');
+        size = defaultBannerSize;
       }
-
-      setState(() {});
 
       if (widget.adUnitId == null) {
         return;
       }
 
-      _anchoredAdaptiveAd = BannerAd(
+      await _bannerAd?.dispose();
+      if (!mounted) return;
+
+      _bannerAd = BannerAd(
         adUnitId: adUnitId!,
         size: size!,
         request: widget.adRequest,
         listener: BannerAdListener(
           onAdLoaded: (Ad ad) {
-            printDebug('$ad loaded: ${ad.responseInfo}');
+            printDebug('onAdLoaded($adUnitId): ${ad.responseInfo}');
             setState(() {
               // When the ad is loaded, get the ad size and use it to set
               // the height of the ad container.
-              _anchoredAdaptiveAd = ad as BannerAd;
+              _bannerAd = ad as BannerAd;
               _isLoaded = true;
             });
           },
           onAdFailedToLoad: (Ad ad, LoadAdError error) {
-            printDebug('Anchored adaptive banner failedToLoad: $error');
+            printDebug('onAdFailedToLoad($adUnitId): $error');
             widget.onAdFailedToLoad?.call(
                 DateTime.now().toUtc().millisecondsSinceEpoch,
                 ad.adUnitId,
@@ -132,15 +117,13 @@ class _GoodBannerAdaptiveAnchoredState
             ad.dispose();
           },
           onAdImpression: (Ad ad) {
-            printDebug('impression: ${ad.responseInfo}');
+            printDebug('onAdImpression($adUnitId): ${ad.responseInfo}');
             widget.onAdImpression?.call(
                 DateTime.now().toUtc().millisecondsSinceEpoch, adUnitId!);
           },
         ),
       );
-      await setLastImpressions(
-          adUnitId!, DateTime.now().millisecondsSinceEpoch);
-      return _anchoredAdaptiveAd!.load();
+      await _bannerAd!.load();
     } catch (e, s) {
       printError('loadAd', e, s);
       widget.onLoadAdError?.call(e, s);
@@ -149,29 +132,39 @@ class _GoodBannerAdaptiveAnchoredState
 
   @override
   Widget build(BuildContext context) {
-    if (_anchoredAdaptiveAd != null && _isLoaded) {
-      return SizedBox(
-        width: _anchoredAdaptiveAd!.size.width.toDouble(),
-        height: _anchoredAdaptiveAd!.size.height.toDouble(),
-        child: AdWidget(ad: _anchoredAdaptiveAd!),
-      );
-    } else {
-      return SizedBox(
-        width: defaultBannerSize?.width.toDouble(),
-        height: defaultBannerSize?.height.toDouble(),
-        child: Center(
-          child: Text(
-            'Ads placeholder',
-            style: TextStyle(color: widget.adsPlaceholderColor),
+    return OrientationBuilder(
+      builder: (context, orientation) {
+        if (_currentOrientation == orientation &&
+            _bannerAd != null &&
+            _isLoaded) {
+          return SizedBox(
+            width: _bannerAd!.size.width.toDouble(),
+            height: _bannerAd!.size.height.toDouble(),
+            child: AdWidget(ad: _bannerAd!),
+          );
+        }
+        // Reload the ad if the orientation changes.
+        if (_currentOrientation != orientation) {
+          _currentOrientation = orientation;
+          _loadAd();
+        }
+        return SizedBox(
+          width: defaultBannerSize?.width.toDouble(),
+          height: defaultBannerSize?.height.toDouble(),
+          child: Center(
+            child: Text(
+              'AD will be here.',
+              style: TextStyle(color: widget.adsPlaceholderColor),
+            ),
           ),
-        ),
-      );
-    }
+        );
+      },
+    );
   }
 
   @override
   void dispose() {
+    _bannerAd?.dispose();
     super.dispose();
-    _anchoredAdaptiveAd?.dispose();
   }
 }
